+++ types/index.ts
@@ -1,0 +1,17 @@
+/**
+ * Central exports for all application types
+ */
+
+// Core data types
+export type { PriceArea, PriceEntry, ThemeMode } from "./data";
+
+// Chart types
+export type { BarDataPoint, ChartConfig, ChartTheme } from "./chart";
+
+// UI types
+export type {
+  ColorScheme,
+  PriceAreaOption,
+  TabConfig,
+  ThemeOption,
+} from "./ui";
